+++ java-impl/netty/src/main/java/com/mawen/learn/redis/basic/command/zset/SortedSetRangeByScoreCommand.java
@@ -41,14 +41,14 @@ public class SortedSetRangeByScoreCommand implements ICommand {
 	public void execute(IDatabase db, IRequest request, IResponse response) {
 		try {
 			DatabaseValue value = db.getOrDefault(request.getParam(0), zset());
-			NavigableSet<Map.Entry<Float, String>> set = value.getValue();
+			NavigableSet<Map.Entry<Double, String>> set = value.getValue();
 
 			float from = parseRange(request.getParam(1));
 			float to = parseRange(request.getParam(2));
 
 			Options options = parseOptions(request);
 
-			Set<Map.Entry<Float, String>> range = set.subSet(
+			Set<Map.Entry<Double, String>> range = set.subSet(
 					score(from, EMPTY_STRING), inclusive(request.getParam(1)),
 					score(to, EMPTY_STRING), inclusive(request.getParam(2)));
