+++ java-impl/netty/src/main/java/com/mawen/learn/redis/basic/command/zset/SortedSetRemoveCommand.java
@@ -33,7 +33,7 @@ public class SortedSetRemoveCommand implements ICommand {
 		List<String> removed = new LinkedList<>();
 
 		db.merge(request.getParam(0), zset(), (oldValue, newValue) -> {
-			Set<Map.Entry<Float, String>> merge = new SortedSet();
+			Set<Map.Entry<Double, String>> merge = new SortedSet();
 			merge.addAll(oldValue.getValue());
 			for (String item : items) {
 				if (merge.remove(item)) {
